requirements.txt
Django==1.11.6
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
PyJWT==1.5.3
pytz==2017.2
